+++ tests/testthat/test-attributes.R
@@ -68,6 +68,17 @@ test_that('number_format works on cells with multiple numbers', {
   number_format(ht)[1, 1] <- list(function(x) ifelse(x > 0, '+', '-'))
 })
 
+
+test_that('number_format does not apply to exponents in scientific notation', {
+  ht <- huxtable(c("1.12e3", "1.12E3", "1.12A3"))
+  number_format(ht) <- 4
+  expect_equivalent(huxtable:::clean_contents(ht, 'latex')[1, 1], "1.1200e3")
+  expect_equivalent(huxtable:::clean_contents(ht, 'latex')[2, 1], "1.1200E3")
+  # the next is not scientific notation so both numbers should be affected
+  expect_equivalent(huxtable:::clean_contents(ht, 'latex')[3, 1], "1.1200A3.0000")
+})
+
+
 test_that('Can combine numbers and strings in padding', {
   ht <- huxtable(a = 1, b = 1)
   left_padding(ht)[1, 1] <- '10pt'